Restructured the init.js file
