+++ assets/init.js
@@ -5,11 +5,11 @@ import {
   SanMovesTable
 } from "https://cdn.jsdelivr.net/npm/@chesslablab/jsblab@0.0.9/src/index.min.js";
 import { INPUT_EVENT_TYPE, COLOR, Chessboard, BORDER_TYPE } from "cm-chessboard";
-import { Accessibility } from "../vendor/cm-chessboard/src/extensions/accessibility/Accessibility.js";
-import { MARKER_TYPE, Markers } from "../vendor/cm-chessboard/src/extensions/markers/Markers.js";
-import { FEN } from "../vendor/cm-chessboard/src/model/Position.js";
-import { PromotionDialog } from "../vendor/cm-chessboard/src/extensions/promotion-dialog/PromotionDialog.js";
-import ChesslaBlabWebSocket from '../ChesslaBlabWebSocket.js';
+import { Accessibility } from "./vendor/cm-chessboard/src/extensions/accessibility/Accessibility.js";
+import { MARKER_TYPE, Markers } from "./vendor/cm-chessboard/src/extensions/markers/Markers.js";
+import { FEN } from "./vendor/cm-chessboard/src/model/Position.js";
+import { PromotionDialog } from "./vendor/cm-chessboard/src/extensions/promotion-dialog/PromotionDialog.js";
+import ChesslaBlabWebSocket from './ChesslaBlabWebSocket.js';
 
 const inputHandler = (event) => {
   if (event.type === INPUT_EVENT_TYPE.movingOverSquare) {
