+++ assets/js/chessboard/classical/index.js
@@ -1,4 +1,4 @@
-import { ws } from '../../init.js';
+import { ws } from '../../../init.js';
 
 localStorage.clear();
 await ws.connect();
